+++ _docs/index.md
@@ -4,10 +4,8 @@
 
 Currently all functionality is accessed with one import:
 ```scala
-import regexglob.RegexGlobbing.r
+import stringmatching.regex.Interpolators.r
 ```
-
-> the exact namespace for the final library will likely change.
 
 A simple example can be illustrated here, where we parse some basic text representing a sequence of integers, delimited by `"["` and `"]"`, and separated by `", "`.
 In one pattern, you can extract a typed value `xs: IndexedSeq[Int]` as follows:
