+++ src/main/test/example/RegexSuite.scala
@@ -1,6 +1,6 @@
-package regexglob
+package example
 
-import RegexGlobbing.r
+import stringmatching.regex.Interpolators.r
 
 import scala.compiletime.testing.typeCheckErrors
